Record healthy periodic healthcheck as appinsight event
This makes it easier to set up automated alerting on it via
Azure Log Analytics

https://trello.com/c/Z67sZszC/135-application-healthchecks
